votes type changed to Integer

+++ src/main/java/com/example/sep6backendv1/Service/IMovieService.java
@@ -16,5 +16,5 @@ public interface IMovieService {
     List<Movie> getMoviesByStar(int personId);
 
     double getRatingById(int id);
-    double getVotesById(int id);
+    Integer getVotesById(int id);
 }
